refactor: pull out use styles to clean up and logically separate code
